added Maven generated site report, sources and javadocs for all submodules
