+++ alexandria-cli/src/test/java/com/github/macgregor/alexandria/cli/ConvertCommandTest.java
@@ -1,14 +1,17 @@
-package com.github.macgregor.alexandria;
+package com.github.macgregor.alexandria.cli;
 
+import com.github.macgregor.alexandria.Alexandria;
+import com.github.macgregor.alexandria.Context;
 import org.junit.Test;
+import org.mockito.Mockito;
 
 import static org.mockito.Mockito.*;
 
-public class SyncCommandTest {
+public class ConvertCommandTest {
 
     @Test
-    public void testSyncIsCalled() throws Exception {
-        SyncCommand testCommand = spy(new SyncCommand());
+    public void testConvertIsCalled() throws Exception {
+        ConvertCommand testCommand = Mockito.spy(new ConvertCommand());
         Alexandria alexandria = spy(new Alexandria());
         Context context = spy(new Context());
         alexandria.context(context);
@@ -17,6 +20,6 @@ public class SyncCommandTest {
         doReturn(alexandria).when(alexandria).syncWithRemote();
         doReturn(alexandria).when(testCommand).alexandria();
         testCommand.call();
-        verify(alexandria, times(1)).syncWithRemote();
+        verify(alexandria, times(1)).convert();
     }
 }
